Fix table sorting in Manage Accounts page. #1
I removed Prototype.js, which was shadowing jQuery, and then made a call
to the `sorttables` function, which adds listeners on the table headings
to toggle sorting.

Prototype.js doesn't look like it has a place in our code, from what I
can tell. Correct me if I'm wrong. I'll be attempting to remove it in
favor of jQuery whenever I get the chance. Until it's "safe," I'll be
preferring `jQuery(...)` over `$(...)`, even though this is generally
not the best style.

I also disabled the functionality to float table headers at the top of
the screen, because it was causing errors and the current "solution" was
really hacky. I might look into enabling this feature later.