Sort all columns with the sortable class.
Fixes #108.

I'm assuming that any tables that include the class="sortable" expect
their contents to be sorted. The way to do this is

  a) include the sorttable.js file
  b) call the function `sorttables();` once the DOM has loaded

I've added step (b) to the sorttable.js file, so that it is done
automatically on all pages that include the JavaScript file.

sorttables.js is an external library, so if we ever update it, this will
get overwritten. However, it is much simpler than pecking and hunting
through the source code to include the three lines required to sort
columns in all the files that require it.