Annotations for PDF submission
This is the initial implementation for a system that will be used by theory
classes.

1. Let's us display PDFs on a page using PDF.js by mozilla
2. Let's the instructors add annotations on a page, similar to the code
   annotations
3. Let's the students view this feedback.
4. Let's both students and instructors download an annotated version of the
   submission as PDF (using Prawn library)

Javascript implementation is a little messy since I built it upon the annotation
system that we use for code annotations. Planning to switch to a good and simple
JS library like Backbone in the future.