Ignore unnecessary MySQL DNS errors
The error message was something like "IP Address 172.x.x.x.x could not
be resolved"

I have a hunch that it was due to the way Docker assigns IP addresses to
containers. I might not have been linking the db container to the web
container correctly, but this worked to shush the issues.